pyproject.toml
[project]
name = "dataset-merger"
version = "0.1.0"
description = "Merge object detection CSV datasets from multiple annotators"
requires-python = ">=3.10"
dependencies = [
    "pandas>=1.3",
    "numpy>=1.21",
    "click>=8.0",
    "scikit-learn>=1.0",
    "pyyaml>=5.4",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
dataset-merger = "merge:main"

[tool.uv]
dev-dependencies = ["pytest>=7.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
